Add v8 nodeinfo command to print out basic information about the node process.
This change adds the nodeinfo command to llscan.cc/.h as it requires a heap scan before it can run. (The results of which are cached and shared with the existing findjsobject and findjsinstances commands.)

It adds accessor functions for retrieving the values of properties within JavaScript
objects to llv8.cc/.h rather than just formatting them as strings. This allows us to examine both the process object and navigate from it to objects referenced from its properties.